refactor: move main logic to `handler` file
