+++ src/next/index.ts
@@ -4,7 +4,7 @@ import {
   NextApiResponse,
 } from "next"
 import { NextAuthOptions, Session } from ".."
-import { IncomingRequest, NextAuthHandler } from "../server"
+import { IncomingRequest, NextAuthHandler } from "../server/handler"
 import extendRes from "../server/lib/extend-res"
 import { set as setCookie } from "../server/lib/cookie"
 import logger, { setLogger } from "../lib/logger"
